A complete rework of our math library: - Changed the API of matrix and vector classes; - Size of Vector2 is now 8 bytes instead of 16 as it no longer uses SSE. Probably SSE is not needed for 2-dimensional vectors. Also I made this class a template, so it can be used with integers, not only floats; - Added PackedVector3F, a 3-dimensional vector that is not aligned to the size of an SSE register; - Got rid of some legacy stuff that is not used anywhere: Half (added like 5 years ago for no particular reason) and component classes for the legacy ECS; - Enabled /fp:fast as an experiment. Presumably we'll get enough precision; - And made some minor changes and refactoring along the way...

+++ FerrumModules/Graphics/RHI/Private/Graphics/RHI/Vulkan/CommandQueue.cpp
@@ -75,7 +75,7 @@ namespace FE::Graphics::Vulkan
         VkSubmitInfo submitInfo{};
         submitInfo.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO;
         submitInfo.pCommandBuffers = commandBuffers.data();
-        submitInfo.commandBufferCount = static_cast<uint32_t>(commandBuffers.size());
+        submitInfo.commandBufferCount = commandBuffers.size();
         submitInfo.pWaitDstStageMask = &kWaitDstFlags;
 
         FE_VK_ASSERT(vkQueueSubmit(m_queue, 1, &submitInfo, VK_NULL_HANDLE));
